♻️ global navbar
