+++ src/App.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import "./App.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Home from "./pages";
@@ -8,13 +8,27 @@ import FounderForm from "./pages/founderForm";
 import InvestorForm from "./pages/investorForm";
 import Onboarding from "./pages/onboarding";
 import AuthenticationGuard from "./components/Common/AuthenticationGuard";
+import Navbar from "./components/Landing/Navbar";
+import Sidebar from "./components/Landing/Sidebar";
 
 function App() {
   const { isAuthenticated } = useAuth0();
+  const [isOpen, setIsOpen] = useState(false);
+
+  const toggle = () => {
+    setIsOpen(!isOpen);
+  };
+
   return (
     <Router>
+      <Sidebar isOpen={isOpen} toggle={toggle} />
+      <Navbar toggle={toggle} />
       <Routes>
-        <Route path="/" element={isAuthenticated ? <Onboarding /> : <Home />} exact />
+        <Route
+          path="/"
+          element={isAuthenticated ? <Onboarding /> : <Home />}
+          exact
+        />
         <Route
           path="/investor-matching"
           element={<AuthenticationGuard component={InvestorForm} />}
