+++ src/components/Onboarding/OnboardingElements.js
@@ -1,20 +1,15 @@
-import styled from 'styled-components'
-
-export const OnboardingContainer = styled.div`
-  background: #0c0c0c;
-  padding-top: 80px;
-`
+import styled from "styled-components";
 
 export const OnboardingStepContainer = styled.div`
   display: flex;
   height: 600px;
   background: white;
-`
+`;
 
 export const OnboardingStepLeftSection = styled.div`
   width: 60%;
   padding: 5%;
-`
+`;
 
 export const OnboardingStepRightSection = styled.div`
   background: #01bf71;
@@ -23,11 +18,11 @@ export const OnboardingStepRightSection = styled.div`
   flex-direction: column;
   justify-content: center;
   padding-left: 5%;
-`
+`;
 
 export const OnboardingCardContainer = styled.div`
   padding: 10% 0;
   display: flex;
   justify-content: space-between;
   gap: 10%;
-`
+`;
