+++ src/pages/index.js
@@ -1,34 +1,23 @@
-import React, { useState } from 'react'
-
-import HeroSection from '../components/Landing/HeroSection'
-import InfoSection from '../components/Landing/infoSection'
+import React from "react";
+import HeroSection from "../components/Landing/HeroSection";
+import InfoSection from "../components/Landing/infoSection";
 import {
   homeObjOne,
   homeObjTwo,
   homeObjThree,
-} from '../components/Landing/infoSection/Data'
-import Navbar from '../components/Landing/Navbar'
-import Services from '../components/Landing/Services'
-import Sidebar from '../components/Landing/Sidebar'
+} from "../components/Landing/infoSection/Data";
+import Services from "../components/Landing/Services";
 
 const Home = () => {
-  const [isOpen, setIsOpen] = useState(false)
-
-  const toggle = () => {
-    setIsOpen(!isOpen)
-  }
-
   return (
     <>
-      <Sidebar isOpen={isOpen} toggle={toggle} />
-      <Navbar toggle={toggle} />
       <HeroSection />
       <InfoSection {...homeObjOne} />
       <InfoSection {...homeObjTwo} />
       <Services />
       <InfoSection {...homeObjThree} />
     </>
-  )
-}
+  );
+};
 
-export default Home
+export default Home;
